Parameterize list of components
This PR extracts virtually all logic from the cmd/otelcontribcol
package into a new internal/otelcontribcore package. What was previously
the main func now requires a ComponentFunc, which provides the list of
components with which the collector should be built. The actual collector
binary is unaffected, but the testbed can easily build a collector with
an alternate (limited) set of components.
