+++ Makefile
@@ -40,7 +40,7 @@ all-modules:
 all: common gotest otelcontribcol otelcontribcol-unstable
 
 .PHONY: e2e-test
-e2e-test: otelcontribcol otelcontribcol-unstable
+e2e-test: otelcontribcol otelcontribcol-unstable otelcontribcol-testbed
 	$(MAKE) -C testbed run-tests
 
 .PHONY: unit-tests-with-cover
@@ -216,6 +216,12 @@ otelcontribcol:
 otelcontribcol-unstable:
 	GO111MODULE=on CGO_ENABLED=0 $(GOCMD) build -trimpath -o ./bin/otelcontribcol_unstable_$(GOOS)_$(GOARCH)$(EXTENSION) \
 		$(BUILD_INFO) -tags $(GO_BUILD_TAGS),enable_unstable ./cmd/otelcontribcol
+
+# Build the Collector executable, with only components used in testbed.
+.PHONY: otelcontribcol-testbed
+otelcontribcol-testbed:
+	GO111MODULE=on CGO_ENABLED=0 $(GOCMD) build -trimpath -o ./bin/otelcontribcol_testbed_$(GOOS)_$(GOARCH)$(EXTENSION) \
+		$(BUILD_INFO) -tags $(GO_BUILD_TAGS),testbed ./cmd/otelcontribcol
 
 .PHONY: otelcontribcol-all-sys
 otelcontribcol-all-sys: otelcontribcol-darwin_amd64 otelcontribcol-darwin_arm64 otelcontribcol-linux_amd64 otelcontribcol-linux_arm64 otelcontribcol-windows_amd64
