+++ cmd/otelcontribcol/main.go
@@ -15,41 +15,17 @@
 // Program otelcontribcol is an extension to the OpenTelemetry Collector
 // that includes additional components, some vendor-specific, contributed
 // from the wider community.
+
+//go:build !testbed
+// +build !testbed
+
 package main
 
 import (
-	"fmt"
-	"log"
-
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/service"
-
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/components"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/version"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/otelcontribcore"
 )
 
 func main() {
-	factories, err := components.Components()
-	if err != nil {
-		log.Fatalf("failed to build components: %v", err)
-	}
-
-	info := component.BuildInfo{
-		Command:     "otelcontribcol",
-		Description: "OpenTelemetry Collector Contrib",
-		Version:     version.Version,
-	}
-
-	if err = run(service.CollectorSettings{BuildInfo: info, Factories: factories}); err != nil {
-		log.Fatal(err)
-	}
-}
-
-func runInteractive(params service.CollectorSettings) error {
-	cmd := service.NewCommand(params)
-	if err := cmd.Execute(); err != nil {
-		return fmt.Errorf("collector server run finished with error: %w", err)
-	}
-
-	return nil
+	otelcontribcore.RunWithComponents(components.Components)
 }
