+++ cmd/otelcontribcol/main_testbed.go
@@ -1,10 +1,10 @@
-// Copyright The OpenTelemetry Authors
+// Copyright 2019 OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//       http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testbed // import "github.com/open-telemetry/opentelemetry-collector-contrib/testbed/testbed"
+// Program otelcontribcol is an extension to the OpenTelemetry Collector
+// that includes additional components, some vendor-specific, contributed
+// from the wider community.
+
+//go:build testbed
+// +build testbed
+
+package main
 
 import (
 	"go.opentelemetry.io/collector/component"
@@ -29,10 +36,15 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/jaegerexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/opencensusexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/zipkinexporter"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/otelcontribcore"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/jaegerreceiver"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/opencensusreceiver"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/zipkinreceiver"
 )
+
+func main() {
+	otelcontribcore.RunWithComponents(Components)
+}
 
 // Components returns the set of components for tests
 func Components() (
